requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==0.58.2
djangorestframework==3.16.0
drf-nested-routers==0.94.2
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
oauthlib==3.3.1
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.27.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
